Wrap error returned from multiClientSubmit
